v2/backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
beautifulsoup4
